Correct the time everytime when pausing in Audio. This will flush the buffer, but is essential for time precision.
Replace the space key listener with a new method #playSwitch, the original code was to simulate the playPause button click, but focusing back and forth between the button and the timeline wastes too much time, which is why the pointer skipping bug happens. Now the code go more straight-forward, which shrink the time and solve the bug.